+++ esp_video/examples/common_components/example_video_common/example_encoder.c
@@ -85,7 +85,7 @@ esp_err_t example_encoder_init(example_encoder_config_t *config, example_encoder
         jpeg_enc_config.sub_sample = JPEG_DOWN_SAMPLING_YUV444;
         jpeg_enc_input_src_size = config->width * config->height * 3;
         break;
-    case V4L2_PIX_FMT_YUV422P:
+    case V4L2_PIX_FMT_UYVY:
         jpeg_enc_config.src_type = JPEG_ENCODE_IN_FORMAT_YUV422;
         jpeg_enc_config.sub_sample = JPEG_DOWN_SAMPLING_YUV422;
         jpeg_enc_input_src_size = config->width * config->height * 2;